Consolidating into fewer files.
Cutting out whatever I can get away with.
