--- std/src/main/scala/lowlevel/CircularBuffer.scala
@@ -1,31 +1,0 @@
-package psp
-package std
-package lowlevel
-
-import api._
-
-final class CircularBuffer[A](capacity: Precise) extends Direct.DirectImpl[A] with AndThis {
-  assert(!capacity.isZero, "CircularBuffer capacity cannot be 0")
-
-  private[this] def cap: Int            = capacity.intValue
-  private[this] val buffer              = newArray[Any](cap)
-  private[this] var seen                = 0L
-  private[this] def writePointer: Int   = (seen % cap).safeInt
-  private[this] def readPointer         = if (isFull) writePointer else 0
-  private[this] def setHead(x: A): Unit = buffer(writePointer) = x sideEffect (seen += 1)
-
-  @inline def foreach(f: A => Unit): Unit = this foreachIndex (i => f(elemAt(i)))
-
-  def head: A                     = elemAt(0.index)
-  def isFull                      = seen >= cap
-  def elemAt(index: Index): A     = buffer((readPointer + index.getInt) % cap).castTo[A]
-  def size: Precise               = capacity min Size(seen)
-  def ++=(xs: Each[A]): this.type = andThis(xs foreach setHead)
-  def += (x: A): this.type        = andThis(this setHead x)
-  def push(x: A): A               = if (isFull) head sideEffect setHead(x) else abort("push on non-full buffer")
-}
-
-object CircularBuffer {
-  def builder[A](capacity: Precise): Builds[A, CircularBuffer[A]] = Builds(xs => CircularBuffer[A](capacity) ++= xs)
-  def apply[A](capacity: Precise): CircularBuffer[A]              = new CircularBuffer[A](capacity)
-}
